refactor(ipfs): move to 'features' folder
