+++ src/features/ipfs/ipfsSettings/ipfsComponents/utilsComponents.jsx
@@ -1,4 +1,4 @@
-import { OptionSelect } from '../../../components';
+import { OptionSelect } from '../../../../components';
 
 function ContainerKeyValue({ children }) {
   return (
